+++ test-parts/test-possible-routes.js
@@ -3,10 +3,11 @@ const chai = require("chai");
 const expect = chai.expect;
 const validOneWay = require("./possible/valid/one-way");
 const validStartEnd = require("./possible/valid/start-end");
+const validStopCount = require("./possible/valid/stop-count");
 const invalidNodeCriteria = require("./possible/invalid/node-criteria");
-const invalidSearchArgs = require("./possible/invalid/node-criteria");
-const invalidStopDistCriteria = require("./possible/invalid/node-criteria");
-const invalidTemplateCriteria = require("./possible/invalid/node-criteria");
+const invalidSearchArgs = require("./possible/invalid/search-args");
+const invalidStopDistCriteria = require("./possible/invalid/stop-dist-criteria");
+const invalidTemplateCriteria = require("./possible/invalid/template-criteria");
 
 
 function runTests()
@@ -25,6 +26,7 @@ function handleValidGroup()
 	{
 		validOneWay.callTest();
 		validStartEnd.callTests();
+		validStopCount.callTests();
 	});
 }